+++ FMCollectionLayout/Classes/Sections/FMLayoutSingleDynamicAddSection.m
@@ -34,7 +34,7 @@
     } else {
         NSInteger items = [self.collectionView numberOfItemsInSection:self.indexPath.section];
         NSMutableArray *attrs = [self.itemsAttribute mutableCopy];
-        for (int j = self.handleItemStart; j < items; j++) {
+        for (int j = (int)self.handleItemStart; j < items; j++) {
             NSIndexPath *indexPath = [NSIndexPath indexPathForItem:j inSection:self.indexPath.section];
             FMCollectionLayoutAttributes *itemAttr = [FMCollectionLayoutAttributes layoutAttributesForCellWithIndexPath:indexPath];
             CGFloat itemWidth =  self.cellFixedWidth;
